+++ ceammc/ext/src/base/mod_base.cpp
@@ -53,6 +53,7 @@ void setup_base_msg();
 void setup_base_replace();
 void setup_base_sync();
 void setup_base_z_tilde();
+void setup_chrono_datetime();
 void setup_is_data();
 void setup_is_file();
 void setup_load_expr();
@@ -103,6 +104,7 @@ void ceammc_base_setup()
     setup_canvas_current();
     setup_canvas_top();
     setup_ceammc_search();
+    setup_chrono_datetime();
     setup_function_call();
     setup_gain_tilde();
     setup_is_data();